Add an abstract class to allow various addition of software
