+++ src/main/groovy/META-INF/MANIFEST.MF
